platform/machine/azure: add azure cluster
Adds the Azure cluster. For each cluster a resource group will be
created inside of Azure which is removed at the end of the clusters
lifecycle (ensuring resource cleanup).

The Azure API requires that all machines are given an SSH key, if the
test is specifically not providing an SSH key the cluster will generate
a fake key via the platform/util function GenerateFakeKey.