platform/api/aws/ec2: create networking resources if missing
Currently kola will only create a SecurityGroup if it is not found. With
this patch it will also create a VPC attached to the SecurityGroup (as
well as the route table, internet gateway, and subnets). This allows
kola to be ran against a fresh account/region and work out of the box.

This patch also switches the filtering of security groups from using the
`GroupNames` parameter which implies that the security groups being
searched are only in the Default VPC to using a Filter on the
`group-names` attribute.