HACK: work around issue with core user creation in FCOS
There is currently a bug where if the core user is created via Ignition
it is not added to the correct set of groups. This completely blocks
kola from running as the core user is not directly added to the sudoers
file with NOPASSWD but rather inherits it from the sudo group.
Temporarily work around it until a fix lands. Upstream tracking ticket:
https://github.com/coreos/fedora-coreos-config/issues/41