kola: add exit-zero-on-test-failure switch
Adds a new switch to the `kola` command, `--exit-zero-on-test-failure`
which when enabled will cause any failures that occur at the individual
test level to not fail the command. This currently does have an
unfortunate limitation that any failures relating to spinning up
infrastructure for said test (such as machine provisioning on clouds)
failing do not fail the run (even when they may be related to the
configuration being provided). Another notable exception is the cluster
semver check, which spins up machines to determine the OS version under
test.